function/requirements.txt
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cognite-sdk==5.11.1
colorama==0.4.6
comm==0.1.3
confection==0.0.4
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
duckduckgo-search==2.8.6
executing==1.2.0
fastai==2.7.12
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.12.0
Fiona==1.9.3
fonttools==4.39.3
geopandas==0.12.2
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
msal==1.22.0
munch==2.5.0
murmurhash==1.0.9
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
packaging==23.1
pandas==2.0.0
parso==0.8.3
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.6.0
pyparsing==3.0.9
pyproj==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
scikit-learn==1.2.2
scipy==1.10.1
shapely==2.0.1
six==1.16.0
smart-open==6.3.0
sortedcontainers==2.4.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
sympy==1.11.1
thinc==8.1.9
threadpoolctl==3.1.0
torch==2.0.0
torchvision==0.15.1
tornado==6.3
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wasabi==1.1.1
wcwidth==0.2.6
zipp==3.15.0